# api/openApi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Link Shortener
  license:
    name: MIT
paths:
  /link:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShortLinkRequest"
      responses:
        201:
          description: Short Link created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShortLink"
        400:
          description: Bad request. Url has wrong schema or body malformed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        409:
          description: Collision with generated key. Try again
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      deprecated: false
  /l/{short-key}:
    parameters:
      - name: short-key
        in: path
        required: true
        schema:
          type: string
    get:
      responses:
        302:
          description: Link found. Redirecting
          headers:
            Location:
              required: true
              schema:
                type: string
                format: url
        400:
          description: Bad request. Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        404:
          description: Link not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      deprecated: false
components:
  schemas:
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          nullable: false
    ShortLink:
      type: object
      required:
        - key
        - url
      properties:
        key:
          type: string
          nullable: false
        url:
          type: string
          format: url
          nullable: false
    ShortLinkRequest:
      type: object
      required:
        - url
      properties:
        url:
          type: string
          format: url
          nullable: false